+++ packages/google-cloud-discoveryengine/google/cloud/discoveryengine/gapic_version.py
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-__version__ = "0.3.1"  # {x-release-please-version}
+__version__ = "0.4.1"  # {x-release-please-version}