+++ packages/google-maps-mapsplatformdatasets/google/maps/mapsplatformdatasets_v1alpha/services/maps_platform_datasets_v1_alpha/client.py
@@ -605,8 +605,8 @@ class MapsPlatformDatasetsV1AlphaClient(
 
         Args:
             request (Union[google.maps.mapsplatformdatasets_v1alpha.types.UpdateDatasetMetadataRequest, dict]):
-                The request object. Request to update the metadata
-                fields of the dataset.
+                The request object. Request to update the metadata fields
+                of the dataset.
             dataset (google.maps.mapsplatformdatasets_v1alpha.types.Dataset):
                 Required. The dataset to update. The dataset's name is
                 used to identify the dataset to be updated. The name has
@@ -719,8 +719,7 @@ class MapsPlatformDatasetsV1AlphaClient(
 
         Args:
             request (Union[google.maps.mapsplatformdatasets_v1alpha.types.GetDatasetRequest, dict]):
-                The request object. Request to get the specified
-                dataset.
+                The request object. Request to get the specified dataset.
             name (str):
                 Required. Resource name. Can also fetch a specified
                 version projects/{project}/datasets/{dataset_id}