+++ packages/google-maps-routing/google/maps/routing_v2/types/routes_service.py
@@ -22,6 +22,7 @@ from google.protobuf import timestamp_pb2  # type: ignore
 from google.rpc import status_pb2  # type: ignore
 import proto  # type: ignore
 
+from google.maps.routing_v2.types import geocoding_results as gmr_geocoding_results
 from google.maps.routing_v2.types import routing_preference as gmr_routing_preference
 from google.maps.routing_v2.types import fallback_info as gmr_fallback_info
 from google.maps.routing_v2.types import polyline, route
@@ -100,7 +101,9 @@ class ComputeRoutesRequest(proto.Message):
             the request fails.
         compute_alternative_routes (bool):
             Optional. Specifies whether to calculate
-            alternate routes in addition to the route.
+            alternate routes in addition to the route. No
+            alternative routes are returned for requests
+            that have intermediate waypoints.
         route_modifiers (google.maps.routing_v2.types.RouteModifiers):
             Optional. A set of conditions to satisfy that
             affect the way routes are calculated.
@@ -113,19 +116,24 @@ class ComputeRoutesRequest(proto.Message):
             for the list of supported languages. When you don't provide
             this value, the display language is inferred from the
             location of the route request.
+        region_code (str):
+            Optional. The region code, specified as a ccTLD ("top-level
+            domain") two-character value. For more information see
+            https://en.wikipedia.org/wiki/List_of_Internet_top-level_domains#Country_code_top-level_domains
         units (google.maps.routing_v2.types.Units):
             Optional. Specifies the units of measure for the display
             fields. This includes the ``instruction`` field in
-            ``NavigationInstruction``. The units of measure used for the
-            route, leg, step distance, and duration are not affected by
-            this value. If you don't provide this value, then the
-            display units are inferred from the location of the request.
+            [NavigationInstruction][google.maps.routing.v2.NavigationInstruction].
+            The units of measure used for the route, leg, step distance,
+            and duration are not affected by this value. If you don't
+            provide this value, then the display units are inferred from
+            the location of the request.
         requested_reference_routes (MutableSequence[google.maps.routing_v2.types.ComputeRoutesRequest.ReferenceRoute]):
             Optional. Specifies what reference routes to calculate as
             part of the request in addition to the default route. A
             reference route is a route with a different route
-            calculation objective than the default route. For example an
-            FUEL_EFFICIENT reference route calculation takes into
+            calculation objective than the default route. For example a
+            ``FUEL_EFFICIENT`` reference route calculation takes into
             account various parameters that would generate an optimal
             fuel efficient route.
         extra_computations (MutableSequence[google.maps.routing_v2.types.ComputeRoutesRequest.ExtraComputation]):
@@ -224,6 +232,10 @@ class ComputeRoutesRequest(proto.Message):
         proto.STRING,
         number=10,
     )
+    region_code: str = proto.Field(
+        proto.STRING,
+        number=16,
+    )
     units: gmr_units.Units = proto.Field(
         proto.ENUM,
         number=11,
@@ -259,6 +271,9 @@ class ComputeRoutesResponse(proto.Message):
             is used, this field contains detailed info about
             the fallback response. Otherwise this field is
             unset.
+        geocoding_results (google.maps.routing_v2.types.GeocodingResults):
+            Contains geocoding response info for
+            waypoints specified as addresses.
     """
 
     routes: MutableSequence[route.Route] = proto.RepeatedField(
@@ -270,6 +285,11 @@ class ComputeRoutesResponse(proto.Message):
         proto.MESSAGE,
         number=2,
         message=gmr_fallback_info.FallbackInfo,
+    )
+    geocoding_results: gmr_geocoding_results.GeocodingResults = proto.Field(
+        proto.MESSAGE,
+        number=3,
+        message=gmr_geocoding_results.GeocodingResults,
     )
 
 
@@ -308,6 +328,19 @@ class ComputeRouteMatrixRequest(proto.Message):
             you made the request. If you set this value to a
             time that has already occurred, the request
             fails.
+        language_code (str):
+            Optional. The BCP-47 language code, such as "en-US" or
+            "sr-Latn". For more information, see
+            http://www.unicode.org/reports/tr35/#Unicode_locale_identifier.
+            See `Language
+            Support <https://developers.google.com/maps/faq#languagesupport>`__
+            for the list of supported languages. When you don't provide
+            this value, the display language is inferred from the
+            location of the first origin.
+        region_code (str):
+            Optional. The region code, specified as a ccTLD ("top-level
+            domain") two-character value. For more information see
+            https://en.wikipedia.org/wiki/List_of_Internet_top-level_domains#Country_code_top-level_domains
         extra_computations (MutableSequence[google.maps.routing_v2.types.ComputeRouteMatrixRequest.ExtraComputation]):
             Optional. A list of extra computations which
             may be used to complete the request. Note: These
@@ -354,6 +387,14 @@ class ComputeRouteMatrixRequest(proto.Message):
         proto.MESSAGE,
         number=5,
         message=timestamp_pb2.Timestamp,
+    )
+    language_code: str = proto.Field(
+        proto.STRING,
+        number=6,
+    )
+    region_code: str = proto.Field(
+        proto.STRING,
+        number=9,
     )
     extra_computations: MutableSequence[ExtraComputation] = proto.RepeatedField(
         proto.ENUM,
@@ -428,11 +469,13 @@ class RouteMatrixElement(proto.Message):
             The travel distance of the route, in meters.
         duration (google.protobuf.duration_pb2.Duration):
             The length of time needed to navigate the route. If you set
-            the ``routing_preference`` to ``TRAFFIC_UNAWARE``, then this
-            value is the same as ``static_duration``. If you set the
-            ``routing_preference`` to either ``TRAFFIC_AWARE`` or
-            ``TRAFFIC_AWARE_OPTIMAL``, then this value is calculated
-            taking traffic conditions into account.
+            the
+            [routing_preference][google.maps.routing.v2.ComputeRouteMatrixRequest.routing_preference]
+            to ``TRAFFIC_UNAWARE``, then this value is the same as
+            ``static_duration``. If you set the ``routing_preference``
+            to either ``TRAFFIC_AWARE`` or ``TRAFFIC_AWARE_OPTIMAL``,
+            then this value is calculated taking traffic conditions into
+            account.
         static_duration (google.protobuf.duration_pb2.Duration):
             The duration of traveling through the route
             without taking traffic conditions into